+++ opencga-catalog/src/main/java/org/opencb/opencga/catalog/db/mongodb/CatalogMongoIndividualDBAdaptor.java
@@ -82,7 +82,7 @@ public class CatalogMongoIndividualDBAdaptor extends CatalogMongoDBAdaptor imple
             throw CatalogDBException.idNotFound("Individual", individual.getMotherId());
         }
 
-        int individualId = getNewId();
+        long individualId = getNewId();
 
         individual.setId(individualId);
